Move lesson entry to models
