+++ src/app/presence-control/components/presence-control-header/presence-control-header.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { PresenceControlHeaderComponent } from './presence-control-header.component';
 import { buildTestModuleMetadata } from 'src/spec-helpers';
 import { buildReference } from 'src/spec-builders';
-import { fromLesson } from '../../utils/lesson-entry';
+import { fromLesson } from '../../models/lesson-entry.model';
 
 describe('PresenceControlHeaderComponent', () => {
   let component: PresenceControlHeaderComponent;
