+++ src/app/presence-control/models/lesson-entry.model.ts
@@ -1,6 +1,6 @@
 import { Lesson } from '../../shared/models/lesson.model';
 import { LessonPresence } from '../../shared/models/lesson-presence.model';
-import { lessonsEqual } from './lessons';
+import { lessonsEqual } from '../utils/lessons';
 
 export function lessonsEntryEqual(
   a: Option<LessonEntry>,
